Refactor entities into draggable or not
