+++ lib/core/flame/components/entities/entity_config.dart
@@ -1,11 +1,11 @@
 import 'package:defend_your_flame/core/flame/components/entities/animation_config.dart';
 import 'package:flame/components.dart';
 
-/// WalkingEntityConfig is a class that holds the information used to define
+/// EntityConfig is a class that holds the information used to define
 /// various concrete implementations of the WalkingEntity class.
 /// Given that most will share the same underlying logic, it makes sense to
 /// simple have a differing config and re-use a shared class handler.
-class WalkingEntityConfig {
+class EntityConfig {
   final String entityResourceName;
 
   final Vector2 defaultSize;
@@ -21,7 +21,7 @@ class WalkingEntityConfig {
 
   final bool canBePickedUp;
 
-  WalkingEntityConfig({
+  EntityConfig({
     required this.entityResourceName,
     required this.defaultSize,
     this.defaultScale = 1.0,
