+++ lib/core/flame/managers/entity_manager.dart
@@ -3,7 +3,7 @@ import 'dart:ui';
 
 import 'package:defend_your_flame/core/flame/components/entities/mobs/skeleton.dart';
 import 'package:defend_your_flame/core/flame/components/entities/mobs/slime.dart';
-import 'package:defend_your_flame/core/flame/components/entities/walking_entity.dart';
+import 'package:defend_your_flame/core/flame/components/entities/entity.dart';
 import 'package:defend_your_flame/core/flame/worlds/main_world.dart';
 import 'package:defend_your_flame/helpers/global_vars.dart';
 import 'package:defend_your_flame/helpers/misc_helper.dart';
@@ -12,7 +12,7 @@ import 'package:flame/components.dart';
 class EntityManager extends Component with ParentIsA<MainWorld> {
   // Used to keep a weak reference to the entities, based on their Y position, so we can render them in the correct order.
   // Using this datastructure as it supports O(log n) for insertion and deletion, and O(n) for iteration.
-  final SplayTreeMap<int, List<WalkingEntity>> _entities = SplayTreeMap();
+  final SplayTreeMap<int, List<Entity>> _entities = SplayTreeMap();
 
   bool _spawning = false;
   int _secondsToSpawnOver = 0;
@@ -20,7 +20,7 @@ class EntityManager extends Component with ParentIsA<MainWorld> {
 
   double _timeCounter = 0;
 
-  bool get roundOver => !_spawning && !children.any((element) => element is WalkingEntity && element.isAlive);
+  bool get roundOver => !_spawning && !children.any((element) => element is Entity && element.isAlive);
 
   int get positionXBoundary => parent.castle.position.x.toInt() - 20;
 
@@ -33,7 +33,7 @@ class EntityManager extends Component with ParentIsA<MainWorld> {
     _entities.clear();
 
     for (var element in children) {
-      if (element is WalkingEntity) {
+      if (element is Entity) {
         element.removeFromParent();
       }
     }
@@ -101,7 +101,7 @@ class EntityManager extends Component with ParentIsA<MainWorld> {
   }
 
   // Wrappers so we can track based on the position of the entity, to render them in the correct order
-  _addEntity(WalkingEntity entity) {
+  _addEntity(Entity entity) {
     var key = entity.position.y.toInt() + entity.scaledSize.y.toInt();
     if (!_entities.containsKey(key)) {
       _entities[key] = [];
